# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger MF-Bank",
        "description": "This is a sample server MF-Bank server.",
        "termsOfService": "http://localhost:8000/api/terms/",
        "contact": {
            "email": "fatih.altun00@hotmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/amountOperations": {
            "put": {
                "tags": [
                    "operations"
                ],
                "summary": "Amount Operations",
                "operationId": "amountOperations",
                "requestBody": {
                    "description": "Amount Operations",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Amount"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid value"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Amount": {
                "title": "Amount Operations",
                "description": "Amount",
                "required": [
                    "amount",
                    "received_user_email"
                ],
                "properties": {
                    "amount": {
                        "type": "integer"
                    },
                    "received_user_email": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}